Revert "Listing retrieval"
