+++ TextCrunch.xcodeproj/xcuserdata/philippabst.xcuserdatad/xcschemes/TextCrunch.xcscheme
@@ -28,7 +28,7 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "655DEF911A8415D200931A3B"
+               BlueprintIdentifier = "658FD0091A64789E004D056D"
                BuildableName = "TextCrunchTests.xctest"
                BlueprintName = "TextCrunchTests"
                ReferencedContainer = "container:TextCrunch.xcodeproj">
@@ -46,7 +46,7 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "655DEF911A8415D200931A3B"
+               BlueprintIdentifier = "658FD0091A64789E004D056D"
                BuildableName = "TextCrunchTests.xctest"
                BlueprintName = "TextCrunchTests"
                ReferencedContainer = "container:TextCrunch.xcodeproj">
@@ -82,11 +82,6 @@
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
-         <AdditionalOption
-            key = "NSZombieEnabled"
-            value = "YES"
-            isEnabled = "YES">
-         </AdditionalOption>
       </AdditionalOptions>
    </LaunchAction>
    <ProfileAction
