+++ TextCrunch/ProfileViewController.swift
@@ -1,0 +1,36 @@
+//
+//  ProfileViewController.swift
+//  TextCrunch
+//
+//  Created by George Coomber on 2015-02-02.
+//  Copyright (c) 2015 Text Crunch. All rights reserved.
+//
+
+import UIKit
+
+class ProfileViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Hide the back navigation button in this view
+		self.navigationItem.setHidesBackButton(true, animated: true)
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
